refactor: move database to shared
