+++ shared/database/common/table.go
